+++ framework/model/src/main/java/is/codion/framework/model/DefaultEntitySearchModel.java
@@ -195,7 +195,7 @@ final class DefaultEntitySearchModel implements EntitySearchModel {
     }
 
     return Select.where(createCombinedCondition(conditions))
-            .limit(limit.isNull() ? -1 : limit.get())
+            .limit(limit.get())
             .build();
   }