+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityTableModel.java
@@ -1052,7 +1052,7 @@ public class SwingEntityTableModel implements EntityTableModel<SwingEntityEditMo
     return Select.where(conditionModel.where(Conjunction.AND))
             .having(conditionModel.having(Conjunction.AND))
             .attributes(selectAttributes())
-            .limit(limit().optional().orElse(-1))
+            .limit(limit().get())
             .orderBy(orderBy())
             .build();
   }